implement additional layer for lnkD, lnk2, lnk3, lnKE key roughly and disable treating warnings as errors. added features are as follows:  - add PascalString and UnicodeString structs often used in psd parsing, see also 'ParsePascalString' and 'ParseUnicodeString' in 'PsdParseMicellaneous.h/cpp'  - add some string functions for convenience, see 'ReadStringFromFile' and 'ReadStringFromFileBE' in 'PsdSyncFileUtil.h/cpp'  - add AdditionalLayer related structs for storing additional layer's info(currently valid for 'lnk2', 'lnk3', 'lnkD', 'lnkE') and functions

+++ src/Psd/PsdSyncFileUtil.h
@@ -21,7 +21,7 @@ namespace fileUtil
 	inline T ReadFromFile(SyncFileReader& reader);
 
 	/// Read char string from a file.
-	std::string ReadStringFromFile(SyncFileReader& reader, size_t count);
+	std::string ReadStringFromFile(SyncFileReader& reader, uint32_t count);
 
 	/// Reads built-in data types from a file, assuming they are stored as big-endian data.
 	/// The read value is automatically converted to the native endianness.
@@ -29,7 +29,7 @@ namespace fileUtil
 	inline T ReadFromFileBE(SyncFileReader& reader);
 
 	/// Read char string from a file, assuming they are stored as big-endian data.
-	std::string ReadStringFromFileBE(SyncFileReader& reader, size_t count);
+	std::string ReadStringFromFileBE(SyncFileReader& reader, uint32_t count);
 
 	/// Writes built-in data types to a file.
 	template <typename T>
